requirements.txt
pip==18.0
apt-xapian-index==0.45
apturl==0.4.1ubuntu4
astroid==1.6.5
backports.functools-lru-cache==1.5
BeautifulSoup==3.2.1
bob==5.0.0
ccsm==0.9.11.3
chardet==2.0.1
cloudpickle==0.5.6
colorama==0.2.5
command-not-found==0.3
compizconfig-python==0.9.11.3
configglue==1.1.2
configobj==4.7.2
configparser==3.5.0
cycler==0.10.0
Cython==0.20.1.post0
dask==0.19.2
debtagshw==0.1
decorator==4.3.0
defer==1.0.6
dirspec==13.10
Django==1.11.16
djangorestframework==3.8.2
dlib==19.15.0
dnspython==1.11.1
duplicity==0.6.23
enum34==1.1.6
feedparser==5.1.3
futures==3.2.0
gunicorn==19.9.0
gyp==0.1
html5lib==0.999
httplib2==0.8
ibus-tegaki==0.3.1
imutils==0.5.1
ipython==1.2.1
isort==4.3.4
Jinja2==2.7.2
Kivy==1.7.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lockfile==0.8
logilab-common==0.61.0
lxml==3.3.3
Mako==0.9.1
MarkupSafe==0.18
matplotlib==2.2.3
mccabe==0.6.1
mss==3.2.1
nemo-emblems==0.0.1
networkx==2.2
numpy==1.15.1
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
opencv-contrib-python==3.4.2.17
paho-mqtt==1.4.0
PAM==0.4.2
paramiko==1.10.1
PassportEye==1.2.2
pdfminer==20140328
pexpect==3.1
Pillow==2.9.0
piston-mini-client==0.7.5
protobuf==2.5.0
psd-tools==0.10
psdtojpg==0.1.0
PyAutoGUI==0.9.38
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygame==1.9.1release
pygobject==3.12.0
pyinotify==0.9.4
pylint==1.9.3
PyMsgBox==1.0.6
pyocr==0.5.2
pyOpenSSL==0.13
pyparsing==2.2.0
PyScreeze==0.1.18
pyserial==2.6
pysmbc==1.0.14.1
pytesseract==0.2.4
python-apt==0.9.3.5ubuntu2
python-dateutil==2.7.3
python-debian==0.1.21-nmu2ubuntu2
python-xlib==0.14
PyTweening==1.0.3
pytz==2018.5
PyWavelets==1.0.1
pyxdg==0.25
pyzmq==14.0.1
reportlab==3.0
requests==2.2.1
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
sessioninstaller==0.0.0
setuptools==40.2.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.5.2
system-service==0.1.6
tegaki-pygtk==0.3.1
tegaki-python==0.3.1
toolz==0.9.0
tornado==3.1.1
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
urllib3==1.7.1
uTidylib==0.2
virtualenv==1.11.4
wrapt==1.10.11
zinnia-python==0.0.0
zope.interface==4.0.5
